making the distance in meters to int

+++ app/src/main/java/com/example/walkapp/models/Performance.kt
@@ -2,7 +2,7 @@ package com.example.walkapp.models
 
 //TODO: Da para adicionar um que mostra o tempo total de caminhada hoje e na semana, por causa da OMS
 class Performance(
-    val distanceTotal: Double,
+    val distanceTotal: Int,
     val distanceLast7Days: List<DistanceDay>,
     val distanceLast12Months: List<DistanceMonth>
 ) {
@@ -17,16 +17,16 @@ class Performance(
     companion object{
         fun mapToPerformance(data: Map<String, Any>): Performance{
             return Performance(
-                distanceTotal = (data["distanceTotal"] as Number).toDouble(),
+                distanceTotal = (data["distanceTotal"] as Number).toInt(),
                 distanceLast7Days = (data["distanceLast7Days"] as List<Map<String, Any>>).map { dayData ->
                     DistanceDay(
-                        distance = (dayData["distance"] as Number).toDouble(),
+                        distance = (dayData["distance"] as Number).toInt(),
                         day = dayData["day"] as String
                     )
                 },
                 distanceLast12Months = (data["distanceLast12Months"] as List<Map<String, Any>>).map { monthData ->
                     DistanceMonth(
-                        distance = (monthData["distance"] as Number).toDouble(),
+                        distance = (monthData["distance"] as Number).toInt(),
                         month = monthData["month"] as String
                     )
                 }
@@ -35,5 +35,5 @@ class Performance(
     }
 }
 
-data class DistanceDay(val distance: Double, val day: String)
-data class DistanceMonth(val distance: Double, val month: String)
+data class DistanceDay(val distance: Int, val day: String)
+data class DistanceMonth(val distance: Int, val month: String)
